+++ ftag/tests/hdf5/test_h5reader.py
@@ -106,6 +106,10 @@ def test_estimate_available_jets(batch_size, num_jets):
     assert estimated_num_jets <= actual_num_jets
     assert estimated_num_jets > 0.95 * actual_num_jets
 
+    # check that the estimate_available_jets function returns the same
+    # number of jets on subsequent calls
+    assert reader.estimate_available_jets(cuts, num=100_000) == estimated_num_jets
+
 
 @pytest.mark.parametrize("equal_jets", [True, False])
 @pytest.mark.parametrize("cuts_list", [["x != -1"], ["x != 1"], ["x == -1"]])
@@ -165,16 +169,14 @@ def test_equal_jets_estimate(equal_jets, cuts_list):
 
 
 def test_reader_transform():
-    fname, f = get_mock_file()
+    fname, _ = get_mock_file()
 
-    transform = Transform(
-        {
-            "jets": {
-                "pt": "pt_new",
-                "silent": "silent",
-            }
+    transform = Transform({
+        "jets": {
+            "pt": "pt_new",
+            "silent": "silent",
         }
-    )
+    })
 
     reader = H5Reader(fname, transform=transform, batch_size=1)
     data = reader.load(num_jets=10)
@@ -184,13 +186,13 @@ def test_reader_transform():
 
 @pytest.fixture
 def singlereader():
-    fname, f = get_mock_file()
+    fname, _ = get_mock_file()
     return H5SingleReader(fname, batch_size=10, do_remove_inf=True)
 
 
 @pytest.fixture
 def reader():
-    fname, f = get_mock_file()
+    fname, _ = get_mock_file()
     return H5Reader(fname, batch_size=10)
 
 
@@ -241,8 +243,8 @@ def test_remove_inf_all_inf_values(singlereader):
 
 
 def test_reader_shapes(reader):
-    assert {"jets": (10,)} == reader.shapes(10)
-    assert {"jets": (10,)} == reader.shapes(10, ["jets"])
+    assert reader.shapes(10) == {"jets": (10,)}
+    assert reader.shapes(10, ["jets"]) == {"jets": (10,)}
 
 
 def test_reader_dtypes(reader):