+++ ftag/tests/test_flavour.py
@@ -5,6 +5,7 @@ import pytest
 from ftag.cuts import Cuts
 from ftag.flavour import (
     Flavour,
+    FlavourContainer,
     Flavours,
     remove_suffix,
 )
@@ -21,6 +22,7 @@ def test_flavour_attributes():
     assert flavour.px == "ptest"
     assert flavour.eff_str == "test_label efficiency"
     assert flavour.rej_str == "test_label rejection"
+    assert flavour.frac_str == "ftest"
     assert str(flavour) == "test"
 
 
@@ -32,6 +34,9 @@ def test_Flavours_iteration():
 def test_Flavours_get_item():
     flavour = Flavours["bjets"]
     assert isinstance(flavour, Flavour)
+    flavour3 = Flavours[Flavours.bjets]
+    assert isinstance(flavour3, Flavour)
+    assert flavour == flavour3
 
 
 def test_Flavours_get_attr():
@@ -41,7 +46,14 @@ def test_Flavours_get_attr():
 
 def test_Flavours_contains():
     assert "bjets" in Flavours
+    assert Flavours.bjets in Flavours
     assert "undefined" not in Flavours
+
+
+def test_Flavours_equals():
+    assert Flavours == FlavourContainer.from_list(Flavours)
+    assert Flavours == [f.name for f in Flavours]
+    assert Flavours != 1
 
 
 def test_Flavours_categories():
@@ -52,6 +64,7 @@ def test_Flavours_categories():
         "xbb-extended",
         "partonic",
         "lepton-decay",
+        "PDGID",
         "isolation",
     ]
     assert Flavours.categories == target
@@ -75,3 +88,26 @@ def test_Flavours_from_cuts():
 def test_remove_suffix():
     assert remove_suffix("test_jets", "jets") == "test_"
     assert remove_suffix("test_jets_test", "jets") == "test_jets_test"
+
+
+def test_backgrounds():
+    bjet_backgrounds = Flavours.backgrounds(Flavours.bjets)
+    print(bjet_backgrounds)
+    assert bjet_backgrounds == FlavourContainer.from_list([
+        Flavours.cjets,
+        Flavours.ujets,
+        Flavours.taujets,
+    ])
+
+    cjet_backgrounds = Flavours.backgrounds(Flavours.cjets)
+    assert cjet_backgrounds == FlavourContainer.from_list([
+        Flavours.bjets,
+        Flavours.ujets,
+        Flavours.taujets,
+    ])
+
+    bjet_background_no_light = Flavours.backgrounds(Flavours.bjets, keep_possible_signals=False)
+    assert bjet_background_no_light == FlavourContainer.from_list([
+        Flavours.cjets,
+        Flavours.taujets,
+    ])