+++ ftag/tests/test_mock.py
@@ -9,7 +9,12 @@ from ftag.mock import JET_VARS, TRACK_VARS, get_mock_file, get_mock_scores
 def test_get_mock_scores():
     labels = np.array([0, 4, 5] * 10)
     scores = get_mock_scores(labels)
-    assert scores.dtype.names == ("MockTagger_pu", "MockTagger_pc", "MockTagger_pb")
+    assert scores.dtype.names == (
+        "MockTagger_pu",
+        "MockTagger_pc",
+        "MockTagger_pb",
+        "MockTagger_ptau",
+    )
     assert scores.shape == (len(labels),)
     assert np.allclose(np.sum(s2u(scores), axis=-1), 1)
 
@@ -25,6 +30,7 @@ def test_get_mock_file():
             "MockTagger_pu",
             "MockTagger_pc",
             "MockTagger_pb",
+            "MockTagger_ptau",
             "MockXbbTagger_phbb",
             "MockXbbTagger_phcc",
             "MockXbbTagger_ptop",
@@ -33,7 +39,8 @@ def test_get_mock_file():
     )
     assert all(jets["pt"] > 0)
     assert all(jets["mass"] <= 50e3)
-    assert all(jets["eta"] >= -3) and all(jets["eta"] <= 3)
+    assert all(jets["eta"] >= -3)
+    assert all(jets["eta"] <= 3)
     assert all(jets["n_truth_promptLepton"] == 0)
 
     # test tracks are correctly generated