+++ ftag/tests/wps/test_discriminant.py
@@ -25,10 +25,57 @@ def test_btag_discriminant():
     tagger = "tagger"
     fc = 0.1
     epsilon = 1e-10
-    disc = btag_discriminant(jets, tagger, fc, epsilon)
+    disc = btag_discriminant(jets, tagger, fc, epsilon=epsilon)
     pb, pc, pu = jets[f"{tagger}_pb"], jets[f"{tagger}_pc"], jets[f"{tagger}_pu"]
     expected = np.log((pb + epsilon) / ((1.0 - fc) * pu + fc * pc + epsilon))
     assert np.allclose(disc, expected)
+
+
+def test_btag_discriminant_inc_tau():
+    jets = np.array(
+        [
+            (0.2, 0.3, 0.9, 0.1),
+            (0.8, 0.5, 0.1, 0.2),
+            (0.6, 0.1, 0.7, 0.3),
+        ],
+        dtype=[
+            ("tagger_pb", "f4"),
+            ("tagger_pc", "f4"),
+            ("tagger_pu", "f4"),
+            ("tagger_ptau", "f4"),
+        ],
+    )
+    tagger = "tagger"
+    fc = 0.1
+
+    epsilon = 1e-10
+    for ftau in (0, 0.2):
+        disc = btag_discriminant(jets, tagger, fc, ftau, epsilon=epsilon)
+        pb, pc, pu, ptau = (jets[f"{tagger}_p{f}"] for f in ("b", "c", "u", "tau"))
+        expected = np.log(
+            (pb + epsilon) / ((1.0 - fc - ftau) * pu + fc * pc + ftau * ptau + epsilon)
+        )
+    assert np.allclose(disc, expected)
+
+
+def test_no_tau_with_ftau():
+    jets = np.array(
+        [
+            (0.2, 0.3, 0.9),
+            (0.8, 0.5, 0.1),
+            (0.6, 0.1, 0.7),
+        ],
+        dtype=[("tagger_pb", "f4"), ("tagger_pc", "f4"), ("tagger_pu", "f4")],
+    )
+    tagger = "tagger"
+    fc = 0.1
+    ftau = 0.2
+    epsilon = 1e-10
+
+    with pytest.raises(ValueError):
+        btag_discriminant(jets, tagger, fc, ftau, epsilon=epsilon)
+    with pytest.raises(ValueError):
+        ctag_discriminant(jets, tagger, fc, ftau, epsilon=epsilon)
 
 
 def test_ctag_discriminant():
@@ -43,7 +90,7 @@ def test_ctag_discriminant():
     tagger = "tagger"
     fb = 0.2
     epsilon = 1e-10
-    disc = ctag_discriminant(jets, tagger, fb, epsilon)
+    disc = ctag_discriminant(jets, tagger, fb, epsilon=epsilon)
     pb, pc, pu = jets[f"{tagger}_pb"], jets[f"{tagger}_pc"], jets[f"{tagger}_pu"]
     expected = np.log((pc + epsilon) / ((1.0 - fb) * pu + fb * pb + epsilon))
     assert np.allclose(disc, expected)
@@ -114,19 +161,76 @@ def test_get_discriminant():
             (0.8, 0.5, 0.1),
             (0.6, 0.1, 0.7),
         ],
-        dtype=[("tagger1_pb", "f4"), ("tagger1_pc", "f4"), ("tagger1_pu", "f4")],
+        dtype=[("tagger_pb", "f4"), ("tagger_pc", "f4"), ("tagger_pu", "f4")],
     )
-    tagger = "tagger1"
+    tagger = "tagger"
     signal = Flavours.bjets
-    disc = get_discriminant(jets, tagger, signal, (0.1,), epsilon=1e-10)
-    expected = btag_discriminant(jets, tagger, fc=0.1, epsilon=1e-10)
+    disc = get_discriminant(jets, tagger, signal, fc=0.1)
+    expected = btag_discriminant(jets, tagger, fc=0.1)
     assert np.allclose(disc, expected)
 
     signal = Flavours.cjets
-    disc = get_discriminant(jets, tagger, signal, (0.2,), epsilon=1e-10)
-    expected = ctag_discriminant(jets, tagger, fb=0.2, epsilon=1e-10)
+    disc = get_discriminant(jets, tagger, signal, fb=0.2)
+    expected = ctag_discriminant(jets, tagger, fb=0.2)
     assert np.allclose(disc, expected)
 
-    # test invalid signal flavour
+    jets = np.array(
+        [
+            (0.2, 0.3, 0.1, 0.4),
+            (0.8, 0.5, 0.2, 0.3),
+            (0.6, 0.1, 0.6, 0.7),
+        ],
+        dtype=[
+            ("tagger_phbb", "f4"),
+            ("tagger_phcc", "f4"),
+            ("tagger_ptop", "f4"),
+            ("tagger_pqcd", "f4"),
+        ],
+    )
+
+    signal = Flavours.hbb
+    disc = get_discriminant(jets, tagger, signal, ftop=0.2, fhcc=0.3)
+    expected = hbb_discriminant(jets, tagger, ftop=0.2, fhcc=0.3)
+    assert np.allclose(disc, expected)
+
+    signal = Flavours.hcc
+    disc = get_discriminant(jets, tagger, signal, ftop=0.2, fhbb=0.3)
+    expected = hcc_discriminant(jets, tagger, ftop=0.2, fhbb=0.3)
+    assert np.allclose(disc, expected)
+
     with pytest.raises(ValueError):
-        get_discriminant(jets, tagger, Flavours.hbb, (0.1,), epsilon=1e-10)
+        get_discriminant(jets, tagger, "blah", ftop=0.2, fhcc=0.3)
+
+
+def test_get_discriminant_tau():
+    jets = np.array(
+        [
+            (0.2, 0.3, 0.5),
+            (0.8, 0.5, 0.1),
+            (0.6, 0.1, 0.7),
+        ],
+        dtype=[("tagger_pb", "f4"), ("tagger_pc", "f4"), ("tagger_pu", "f4")],
+    )
+
+    tagger = "tagger"
+    signal = Flavours.bjets
+    with pytest.raises(ValueError):
+        get_discriminant(jets, tagger, signal, fc=0.1, ftau=0.1)
+
+    jets = np.array(
+        [
+            (0.2, 0.3, 0.5, 0.1),
+            (0.8, 0.5, 0.1, 0.2),
+            (0.6, 0.1, 0.7, 0.3),
+        ],
+        dtype=[
+            ("tagger_pb", "f4"),
+            ("tagger_pc", "f4"),
+            ("tagger_pu", "f4"),
+            ("tagger_ptau", "f4"),
+        ],
+    )
+
+    disc = get_discriminant(jets, tagger, Flavours.bjets, fc=0.1, ftau=0.1)
+    expected = btag_discriminant(jets, tagger, fc=0.1, ftau=0.1)
+    assert np.allclose(disc, expected)